fix(eslint-errors): accordion
