+++ packages/accordion/src/accordion/context.ts
@@ -1,0 +1,4 @@
+import React from 'react';
+import { UseAccordionProps } from './accordion';
+
+export const AccordionContext = React.createContext<UseAccordionProps>({ allowMultipleOpen: false });
